requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy>=2.0.25
requests==2.31.0
jinja2==3.1.2
gunicorn==21.2.0
